forecasting-tools/pyproject.toml
[tool.poetry]
name = "forecasting-tools"
version = "0.2.39"
description = "AI forecasting and research tools to help humans reason about and forecast the future"
authors = ["Benjamin Wilson <mokoresearch@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = ["ai", "forecasting", "research", "metaculus", "prediction", "future", "market", "personality"]
repository = "https://github.com/Metaculus/forecasting-tools"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Natural Language :: English",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.51.0"
tiktoken = ">=0.8.0,<0.10.0"
aiofiles = "^24.1.0"
aiohttp = "^3.9.3"
aiolimiter = "^1.1.0"
asyncio = "^3.4.3"
nest-asyncio = "^1.5.8"
requests = "^2.32.3"
pydantic = "^2.9.2"
python-dotenv = "^1.0.0"
regex = "^2024.11.6"
tenacity = ">=8.0.0,<10.0.0"
pillow = ">=9.0.0,<12.0.0"
transformers = "^4.44.1"
typeguard = "^4.3.0"
scikit-learn = "^1.5.2"
litellm = "^1.59.1"
numpy = ">=1.26.0,<3.0.0"
exceptiongroup = "^1.2.2"
scipy = "^1.12.0"
faker = "^37.0.1"
plotly = ">=5.24.1,<7.0.0"
pandas = "^2.2.3"
streamlit = "^1.20.0"
asknews = ">=0.9.1,<0.12.0"
openai-agents = {extras = ["litellm"], version = "^0.0.14"}
# Personality management dependencies
sqlite3-extras = "^0.0.3"  # Enhanced SQLite functions for telemetry
jinja2 = "^3.1.3"  # Template handling improvements
pytest-benchmark = "^4.0.0"  # For performance benchmarking

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
seaborn = "^0.13.2"
matplotlib = "^3.9.2"
ipykernel = "^6.29.5"
notebook = "^7.2.2"
pytest = "^8.3.2"
pytest-xdist = "^3.3.1"
pytest-mock = "^3.14.0"
pytest-asyncio = ">=0.25.0,<0.27.0"
pypistats = "^1.7.0"
tox = "^4.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
